+++ src/main/java/org/candlepin/subscriptions/subscription/export/ExportSubscriptionListener.java
@@ -18,9 +18,9 @@
  * granted to use or replicate Red Hat trademarks that are incorporated
  * in this software or its documentation.
  */
-package org.candlepin.subscriptions.subscription;
+package org.candlepin.subscriptions.subscription.export;
 
-import static org.candlepin.subscriptions.ApplicationConfiguration.SUBSCRIPTION_EXPORT_QUALIFIER;
+import static org.candlepin.subscriptions.subscription.export.ExportSubscriptionConfiguration.SUBSCRIPTION_EXPORT_QUALIFIER;
 
 import com.redhat.cloud.event.apps.exportservice.v1.Format;
 import com.redhat.cloud.event.apps.exportservice.v1.ResourceRequest;
